Initial version of Russian translation
it was by far the hardest translation for me, since it was the biggest plugin i have ever translated. Of course it has some typos, errors in logic and wrong translations and so on, but i will try to fix all of this. Now i have learned some lessons and i think i can improve my translators skills. Thanks for this opportunity!